backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Vector Database
faiss-cpu==1.7.4
numpy==1.24.3

# Database
aiosqlite==0.19.0

# Markdown Processing
markdown==3.5.2
python-frontmatter==1.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Search
ddgs>=0.1.0  # 新包名 (原 duckduckgo-search)

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utils
tqdm==4.66.1
watchdog==3.0.0
python-dateutil==2.9.0
beautifulsoup4==4.13.3
